requirements.txt
asgiref==3.7.2
Django==5.0.3
django-filter==24.1
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
PyJWT==1.7.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
uuid==1.30
